requirements.txt
albumentations==1.3.1
diffusers==0.28.2
einops==0.8.0
faiss==1.8.0
hydra-core==1.3.2
joblib==1.4.2
ldm==0.1.3
loguru==0.7.2
mat73==0.62
matplotlib==3.9.0
numpy==2.0.0
omegaconf==2.3.0
open_clip_torch==2.24.0
Pillow==10.3.0
Requests==2.32.3
scipy==1.13.1
segment_anything==1.0
setuptools==69.5.1
timm==1.0.7
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.41.2
~orch==2.3.0
